Update documentation across various guides
Updated K3s usage instructions, provided a reference link in Terraform AWS README and enhanced Minio alias setting instructions for on-prem setup. This revision helps in improving clarity and usage instruction precision especially for K3s and on-premises AWS setup. A reference is also added to the AWS README for further information.